refactor: extract data transfer in DTO
