+++ Doppler.PushContact/Repositories/Interfaces/IDomainRepository.cs
@@ -1,11 +1,12 @@
 using Doppler.PushContact.Models;
+using Doppler.PushContact.Models.DTOs;
 using System.Threading.Tasks;
 
 namespace Doppler.PushContact.Repositories.Interfaces
 {
     public interface IDomainRepository
     {
-        Task UpsertAsync(Domain domain);
-        Task<Domain> GetByNameAsync(string name);
+        Task UpsertAsync(DomainDTO domain);
+        Task<DomainDTO> GetByNameAsync(string name);
     }
 }
