+++ Doppler.PushContact/Services/DomainService.cs
@@ -1,4 +1,4 @@
-using Doppler.PushContact.Models;
+using Doppler.PushContact.Models.DTOs;
 using Doppler.PushContact.Repositories.Interfaces;
 using Microsoft.Extensions.Logging;
 using System.Threading.Tasks;
@@ -19,12 +19,12 @@ namespace Doppler.PushContact.Services
             _logger = logger;
         }
 
-        public async Task UpsertAsync(Domain domain)
+        public async Task UpsertAsync(DomainDTO domain)
         {
             await _domainRepository.UpsertAsync(domain);
         }
 
-        public async Task<Domain> GetByNameAsync(string name)
+        public async Task<DomainDTO> GetByNameAsync(string name)
         {
             var domain = await _domainRepository.GetByNameAsync(name);
             return domain;
